[feature] add service layer and fix query

+++ models/order.go
@@ -10,7 +10,7 @@ const (
 
 type Order struct {
 	*gorm.Model
-	VendorId     int64  `json:"vendor_id"`
+	VendorId     uint   `json:"vendor_id"`
 	DeliveryTime int64  `json:"delivery_time"`
 	State        string `json:"state"`
 }
